Restructuring code and implementing hide/show functionality
